feat: moved steps dir components
