+++ app/src/i18n/locales/en/onboarding.json
@@ -29,8 +29,8 @@
   "inventory-year-required": "Please select year",
   "gpc-basic-message": "Only GPC Basic Inventories are supported momentarily. For population values the closest year to the city inventory is recommended, as it will be used for calculations and included in your GPC report.",
   "inventory-goal-required": "Please select inventory goal",
-  "gpc_basic": "GPC BASIC",
-  "gpc_basic_plus": "GPC BASIC+",
+  "gpc-basic": "GPC BASIC",
+  "gpc-basic-plus": "GPC BASIC+",
   "save-button": "Save and Continue",
   "search-city-button": "Search for another City",
   "confirm-button": "Confirm and Continue",
